fix: correcciones de dialecto y datos post-migración Databricks
- fct_transactions: EXTRACT(DOW)::int -> dayofweek()-1 (Spark)
- int_customer_monthly_revenue: eliminar ::numeric (incompatible con Spark ANSI)
- mart_customer_360: ::numeric -> cast(... as double) en agregaciones de revenue
- mart_acquisition_trend: to_char/date_trunc -> date_format/trunc (Spark)
- mart_loan_composition/mart_top_merchants: coalesce en SUM/AVG para NULLs
- flatten_transactions: amount StringType + regexp_replace antes de cast a double
- flatten_loans: principal/outstanding_balance/interest_rate/monthly_payment
  StringType + regexp_replace antes de cast (valores con prefijo $ en JSON)
- dbt test final: PASS=427, WARN=7, ERROR=0 (434 tests totales)

+++ dbt/models/intermediate/int_customer_monthly_revenue.sql
@@ -118,7 +118,7 @@ final as (
 
         -- ---------- Lifetime fee revenue (kept for auditability) ----------
         -- Native: rounded (display-grain, single customer)
-        coalesce(round(f.fee_revenue_lifetime_native::numeric, 2), 0)
+        coalesce(round(f.fee_revenue_lifetime_native, 2), 0)
             as total_fees_paid_lifetime_native,
         -- USD: full precision (downstream AVG over segment requires it)
         coalesce(f.fee_revenue_lifetime_usd, 0)
@@ -128,7 +128,7 @@ final as (
         -- Native: rounded
         coalesce(
             round(
-                (f.fee_revenue_lifetime_native / nullif(c.tenure_months, 0))::numeric,
+                (f.fee_revenue_lifetime_native / nullif(c.tenure_months, 0)),
                 2
             ),
             0
@@ -141,7 +141,7 @@ final as (
 
         -- ---------- Monthly interest revenue ----------
         -- Native: rounded
-        coalesce(round(i.monthly_interest_revenue_native::numeric, 2), 0)
+        coalesce(round(i.monthly_interest_revenue_native, 2), 0)
             as monthly_interest_revenue_native,
         -- USD: full precision
         coalesce(i.monthly_interest_revenue_usd, 0)
@@ -151,11 +151,11 @@ final as (
         -- Native: sum of rounded components
         coalesce(
             round(
-                (f.fee_revenue_lifetime_native / nullif(c.tenure_months, 0))::numeric,
+                (f.fee_revenue_lifetime_native / nullif(c.tenure_months, 0)),
                 2
             ),
             0
-        ) + coalesce(round(i.monthly_interest_revenue_native::numeric, 2), 0)
+        ) + coalesce(round(i.monthly_interest_revenue_native, 2), 0)
             as total_monthly_revenue_native,
         -- USD: sum of full-precision components
         coalesce(
